# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateShipRequest:
    properties:
      configuration_id:
        type: string
      imo:
        type: string
      name:
        type: string
      organization_id:
        type: string
      passenger_capacity:
        type: integer
      status:
        type: string
      total_seats:
        type: integer
      type_ship:
        type: string
      weight_capacity:
        type: number
    required:
    - imo
    - name
    - passenger_capacity
    - status
    - total_seats
    - type_ship
    - weight_capacity
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: detailed error message
        type: string
    type: object
...
  dto.ListShipResponse:
    properties:
      current_page:
        type: integer
      ships:
        items:
          $ref: '#/definitions/dto.ShipListItemResponse'
        type: array
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.ShipListItemResponse:
    properties:
      configuration_id:
        type: string
      id:
        type: string
      image_url:
        items:
          type: string
        type: array
      imo:
        type: string
      name:
        type: string
      organization_id:
        type: string
      passenger_capacity:
        type: integer
      status:
        type: string
      total_seats:
        type: integer
      type_ship:
        type: string
      weight_capacity:
        type: number
    type: object
  dto.ShipResponse:
    properties:
      configuration_id:
        type: string
      id:
        type: string
      image_url:
        items:
          type: string
        type: array
      imo:
        type: string
      name:
        type: string
      organization_id:
        type: string
      passenger_capacity:
        type: integer
      status:
        type: string
      total_seats:
        type: integer
      type_ship:
        type: string
      weight_capacity:
        type: number
    type: object
...
  dto.UpdateShipRequest:
    properties:
      configuration_id:
        type: string
      imo:
        type: string
      name:
        type: string
      organization_id:
        type: string
      passenger_capacity:
        type: integer
      status:
        type: string
      total_seats:
        type: integer
      type_ship:
        type: string
      weight_capacity:
        type: number
    type: object
...
host: localhost:8002
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Service for managing user identities, roles, and organizations.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Uber dos Rios API
  version: v1.0.16
paths:
...
  /ships:
    get:
      consumes:
      - application/json
      description: Retrieves a list of ships with pagination.
      parameters:
      - description: Ship type (BOAT, FERRYBOAT or SHIP)
        in: query
        name: typeShip
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ListShipResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List Ships
      tags:
      - ships
    post:
      consumes:
      - application/json
      description: Registers a new ship with the provided details.
      parameters:
      - description: SHIP Data to Create
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/dto.CreateShipRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ShipResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new ship
      tags:
      - ships
...
  /ships/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a ship by its ID.
      parameters:
      - description: Ship ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID supplied
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Ship not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a ship
      tags:
      - ships
    get:
      consumes:
      - application/json
      description: Retrieves a Ship by their ID.
      parameters:
      - description: Ship type (BOAT, FERRYBOAT or SHIP)
        in: query
        name: typeShip
        type: string
      - description: Ship ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ShipResponse'
        "400":
          description: Invalid SHip ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Ship not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get Ship by ID
      tags:
      - ships
    put:
      consumes:
      - application/json
      description: Updates an existing ship with the provided details.
      parameters:
      - description: Ship ID
        in: path
        name: id
        required: true
        type: string
      - description: Ship Data to Update
        in: body
        name: ship
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateShipRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ShipResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: SHIP not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a ship
      tags:
      - ships
...
schemes:
- http
- https
swagger: "2.0"